refactor: move CurrentUser annotation from api/common to api/auth
CurrentUser was the only file under api/common/security/, and it exists
solely to surface the auth slice's Principal to controllers via
@AuthenticationPrincipal. Its natural owner is auth, not common — the
other slices that use it are consumers, not co-owners.

Moves:
- api/common/security/annotation/CurrentUser -> api/auth/security/annotation/CurrentUser
- 5 controller imports updated (post, comment, meeting, user, file)
- api/common/security/ directory removed (now empty)

No behavior change. docs/ARCHITECTURE.md's rule "common holds only what
two or more slices co-own" is now enforced for this case.
