+++ eventitta-api/src/main/java/com/eventitta/api/auth/security/annotation/CurrentUser.java
@@ -1,4 +1,4 @@
-package com.eventitta.api.common.security.annotation;
+package com.eventitta.api.auth.security.annotation;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
