+++ eventitta-api/src/main/java/com/eventitta/api/meeting/controller/MeetingController.java
@@ -1,6 +1,6 @@
 package com.eventitta.api.meeting.controller;
 
-import com.eventitta.api.common.security.annotation.CurrentUser;
+import com.eventitta.api.auth.security.annotation.CurrentUser;
 import com.eventitta.api.common.response.ApiErrorResponse;
 import com.eventitta.domain.common.response.PageResponse;
 import com.eventitta.domain.meeting.dto.request.MeetingCreateRequest;
